+++ src/model/Vertex.ts
@@ -7,7 +7,7 @@
 module StateJS {
 	/**
 	 * An abstract element within a state machine model that can be the source or target of a transition (states and pseudo states).
-	 * 
+	 *
 	 * Vertex extends the Element class and inherits its public interface.
 	 * @class Vertex
 	 * @augments Element
@@ -18,20 +18,20 @@ module StateJS {
 		 * @member {Region}
 		 */
 		public region: Region;
-		
+
 		/**
 		 * The set of transitions from this vertex.
 		 * @member {Array<Transition>}
 		 */
 		public transitions: Array<Transition> = [];
-	
+
 		/**
 		 * Creates a new instance of the Vertex class within a given parent region.
 		 * @param {string} name The name of the vertex.
 		 * @param {Region} parent The parent region.
 		 */
 		public constructor(name: string, parent: Region);
-		
+
 		/**
 		 * Creates a new instance of the Vertex class within a given parent state.
 		 * Note, this will create the vertex within the parent states default region.
@@ -39,7 +39,7 @@ module StateJS {
 		 * @param {State} parent The parent state.
 		 */
 		public constructor(name: string, parent: State);
-	
+
 		/**
 		 * Creates a new instance of the Vertex class.
 		 * @param {string} name The name of the vertex.
@@ -47,19 +47,19 @@ module StateJS {
 		 */
 		public constructor(name: string, parent: any) {
 			super(name);
-	
+
 			if (parent instanceof Region) {
-				this.region = <Region>parent;
+				this.region = parent;
 			} else if (parent instanceof State) {
-				this.region = (<State>parent).defaultRegion();
+				this.region = parent.defaultRegion();
 			}
-	
+
 			if (this.region) {
 				this.region.vertices.push(this);
 				this.region.getRoot().clean = false;
 			}
 		}
-	
+
 		/**
 		 * Returns the parent element of this vertex.
 		 * @method getParent
@@ -68,17 +68,17 @@ module StateJS {
 		public getParent(): Element {
 			return this.region;
 		}
-		
+
 		// TODO: find a clean way to remove this
 		public isJunction(): boolean {
 			return false;
 		}
-		
+
 		// TODO: find a clean way to remove this
 		public isChoice(): boolean {
 			return false;
 		}
-		
+
 		/**
 		 * Creates a new transition from this vertex.
 		 * Newly created transitions are completion transitions; they will be evaluated after a vertex has been entered if it is deemed to be complete.
@@ -90,10 +90,10 @@ module StateJS {
 		 */
 		public to(target?: Vertex, kind: TransitionKind = TransitionKind.External): Transition {
 			var transition = new Transition(this, target, target ? kind : TransitionKind.Internal);
-	
+
 			this.transitions.push(transition);
 			this.getRoot().clean = false;
-	
+
 			return transition;
 		}
 	}